socket_service_latest/requirements.txt
Flask==1.1.1
flask_cors==3.0.7
kafka-python==2.0.1
Flask-SocketIO==5.0.1
eventlet==0.30.1